+++ pkg/BUILD.bazel
@@ -76,8 +76,6 @@ ALL_TESTS = [
     "//pkg/ccl/multitenantccl/tenantcostserver:tenantcostserver_test",
     "//pkg/ccl/oidcccl:oidcccl_test",
     "//pkg/ccl/partitionccl:partitionccl_test",
-    "//pkg/ccl/pgcryptoccl/pgcryptocipherccl:pgcryptocipherccl_test",
-    "//pkg/ccl/pgcryptoccl:pgcryptoccl_test",
     "//pkg/ccl/schemachangerccl/sctestbackupccl:sctestbackupccl_test",
     "//pkg/ccl/schemachangerccl/sctestbackupmrccl:sctestbackupmrccl_test",
     "//pkg/ccl/schemachangerccl:schemachangerccl_test",
@@ -603,6 +601,8 @@ ALL_TESTS = [
     "//pkg/sql/parser:parser_disallowed_imports_test",
     "//pkg/sql/parser:parser_test",
     "//pkg/sql/parserutils:parserutils_disallowed_imports_test",
+    "//pkg/sql/pgcrypto/pgcryptocipher:pgcryptocipher_test",
+    "//pkg/sql/pgcrypto:pgcrypto_test",
     "//pkg/sql/pgrepl/pgreplparser:pgreplparser_test",
     "//pkg/sql/pgrepl:pgrepl_test",
     "//pkg/sql/pgwire/hba:hba_test",
@@ -1056,10 +1056,6 @@ GO_TARGETS = [
     "//pkg/ccl/oidcccl:oidcccl_test",
     "//pkg/ccl/partitionccl:partitionccl",
     "//pkg/ccl/partitionccl:partitionccl_test",
-    "//pkg/ccl/pgcryptoccl/pgcryptocipherccl:pgcryptocipherccl",
-    "//pkg/ccl/pgcryptoccl/pgcryptocipherccl:pgcryptocipherccl_test",
-    "//pkg/ccl/pgcryptoccl:pgcryptoccl",
-    "//pkg/ccl/pgcryptoccl:pgcryptoccl_test",
     "//pkg/ccl/schemachangerccl/sctestbackupccl:sctestbackupccl_test",
     "//pkg/ccl/schemachangerccl/sctestbackupmrccl:sctestbackupmrccl_test",
     "//pkg/ccl/schemachangerccl:schemachangerccl",
@@ -2358,6 +2354,10 @@ GO_TARGETS = [
     "//pkg/sql/parser:parser",
     "//pkg/sql/parser:parser_test",
     "//pkg/sql/parserutils:parserutils",
+    "//pkg/sql/pgcrypto/pgcryptocipher:pgcryptocipher",
+    "//pkg/sql/pgcrypto/pgcryptocipher:pgcryptocipher_test",
+    "//pkg/sql/pgcrypto:pgcrypto",
+    "//pkg/sql/pgcrypto:pgcrypto_test",
     "//pkg/sql/pgrepl/lsn:lsn",
     "//pkg/sql/pgrepl/lsnutil:lsnutil",
     "//pkg/sql/pgrepl/pgreplparser:pgreplparser",
@@ -2492,7 +2492,6 @@ GO_TARGETS = [
     "//pkg/sql/sem/asof:asof_test",
     "//pkg/sql/sem/builtins/builtinconstants:builtinconstants",
     "//pkg/sql/sem/builtins/builtinsregistry:builtinsregistry",
-    "//pkg/sql/sem/builtins/pgcrypto:pgcrypto",
     "//pkg/sql/sem/builtins/pgformat:pgformat",
     "//pkg/sql/sem/builtins/pgformat:pgformat_test",
     "//pkg/sql/sem/builtins:builtins",
