+++ pkg/ccl/BUILD.bazel
@@ -18,7 +18,6 @@ go_library(
         "//pkg/ccl/multitenantccl",
         "//pkg/ccl/oidcccl",
         "//pkg/ccl/partitionccl",
-        "//pkg/ccl/pgcryptoccl",
         "//pkg/ccl/securityccl/fipsccl",
         "//pkg/ccl/storageccl",
         "//pkg/ccl/utilccl",
