+++ pkg/sql/pgcrypto/BUILD.bazel
@@ -1,32 +1,24 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
-    name = "pgcryptoccl",
-    srcs = ["pgcryptoccl.go"],
-    importpath = "github.com/cockroachdb/cockroach/pkg/ccl/pgcryptoccl",
+    name = "pgcrypto",
+    srcs = ["pgcrypto.go"],
+    importpath = "github.com/cockroachdb/cockroach/pkg/sql/pgcrypto",
     visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/ccl/pgcryptoccl/pgcryptocipherccl",
-        "//pkg/ccl/utilccl",
-        "//pkg/docs",
-        "//pkg/sql/sem/builtins/pgcrypto",
-        "//pkg/sql/sem/eval",
-    ],
+    deps = ["//pkg/sql/pgcrypto/pgcryptocipher"],
 )
 
 go_test(
-    name = "pgcryptoccl_test",
+    name = "pgcrypto_test",
     srcs = [
         "main_test.go",
-        "pgcryptoccl_test.go",
+        "pgcrypto_test.go",
     ],
     deps = [
         "//pkg/base",
-        "//pkg/ccl",
         "//pkg/security/securityassets",
         "//pkg/security/securitytest",
         "//pkg/server",
-        "//pkg/testutils",
         "//pkg/testutils/serverutils",
         "//pkg/testutils/testcluster",
         "//pkg/util/leaktest",
