+++ pkg/sql/pgcrypto/main_test.go
@@ -3,7 +3,7 @@
 // Use of this software is governed by the CockroachDB Software License
 // included in the /LICENSE file.
 
-package pgcryptoccl_test
+package pgcrypto_test
 
 import (
 	"os"
