+++ pkg/sql/pgcrypto/pgcryptocipher/cipher_fuzz.go
@@ -3,7 +3,7 @@
 // Use of this software is governed by the CockroachDB Software License
 // included in the /LICENSE file.
 
-package pgcryptocipherccl
+package pgcryptocipher
 
 import (
 	"crypto/aes"
