+++ pkg/sql/pgcrypto/pgcryptocipher/cipher_test.go
@@ -3,19 +3,19 @@
 // Use of this software is governed by the CockroachDB Software License
 // included in the /LICENSE file.
 
-package pgcryptocipherccl_test
+package pgcryptocipher_test
 
 import (
 	"testing"
 
-	"github.com/cockroachdb/cockroach/pkg/ccl/pgcryptoccl/pgcryptocipherccl"
+	"github.com/cockroachdb/cockroach/pkg/sql/pgcrypto/pgcryptocipher"
 	"github.com/stretchr/testify/require"
 )
 
 func TestEncrypt(t *testing.T) {
-	for name, tc := range pgcryptocipherccl.CipherTestCases {
+	for name, tc := range pgcryptocipher.CipherTestCases {
 		t.Run(name, func(t *testing.T) {
-			res, err := pgcryptocipherccl.Encrypt(tc.Plaintext, tc.Key, tc.Iv, tc.CipherType)
+			res, err := pgcryptocipher.Encrypt(tc.Plaintext, tc.Key, tc.Iv, tc.CipherType)
 			require.NoError(t, err)
 			require.Equal(t, tc.Ciphertext, res)
 		})
@@ -23,9 +23,9 @@ func TestEncrypt(t *testing.T) {
 }
 
 func TestDecrypt(t *testing.T) {
-	for name, tc := range pgcryptocipherccl.CipherTestCases {
+	for name, tc := range pgcryptocipher.CipherTestCases {
 		t.Run(name, func(t *testing.T) {
-			res, err := pgcryptocipherccl.Decrypt(tc.Ciphertext, tc.Key, tc.Iv, tc.CipherType)
+			res, err := pgcryptocipher.Decrypt(tc.Ciphertext, tc.Key, tc.Iv, tc.CipherType)
 			require.NoError(t, err)
 			require.Equal(t, tc.Plaintext, res)
 		})
@@ -33,7 +33,7 @@ func TestDecrypt(t *testing.T) {
 }
 
 func BenchmarkEncrypt(b *testing.B) {
-	for name, tc := range pgcryptocipherccl.CipherTestCases {
+	for name, tc := range pgcryptocipher.CipherTestCases {
 		b.Run(name, func(b *testing.B) {
 			benchmarkEncrypt(b, tc.Plaintext, tc.Key, tc.Iv, tc.CipherType)
 		})
@@ -41,7 +41,7 @@ func BenchmarkEncrypt(b *testing.B) {
 }
 
 func BenchmarkDecrypt(b *testing.B) {
-	for name, tc := range pgcryptocipherccl.CipherTestCases {
+	for name, tc := range pgcryptocipher.CipherTestCases {
 		b.Run(name, func(*testing.B) {
 			benchmarkDecrypt(b, tc.Ciphertext, tc.Key, tc.Iv, tc.CipherType)
 		})
@@ -50,14 +50,14 @@ func BenchmarkDecrypt(b *testing.B) {
 
 func benchmarkEncrypt(b *testing.B, data []byte, key []byte, iv []byte, cipherType string) {
 	for n := 0; n < b.N; n++ {
-		_, err := pgcryptocipherccl.Encrypt(data, key, iv, cipherType)
+		_, err := pgcryptocipher.Encrypt(data, key, iv, cipherType)
 		require.NoError(b, err)
 	}
 }
 
 func benchmarkDecrypt(b *testing.B, data []byte, key []byte, iv []byte, cipherType string) {
 	for n := 0; n < b.N; n++ {
-		_, err := pgcryptocipherccl.Decrypt(data, key, iv, cipherType)
+		_, err := pgcryptocipher.Decrypt(data, key, iv, cipherType)
 		require.NoError(b, err)
 	}
 }
