+++ pkg/sql/sem/builtins/BUILD.bazel
@@ -74,6 +74,7 @@ go_library(
         "//pkg/sql/opt/workloadindexrec",
         "//pkg/sql/parser/statements",
         "//pkg/sql/parserutils",
+        "//pkg/sql/pgcrypto",
         "//pkg/sql/pgwire/pgcode",
         "//pkg/sql/pgwire/pgerror",
         "//pkg/sql/pgwire/pgnotice",
@@ -85,7 +86,6 @@ go_library(
         "//pkg/sql/sem/asof",
         "//pkg/sql/sem/builtins/builtinconstants",
         "//pkg/sql/sem/builtins/builtinsregistry",
-        "//pkg/sql/sem/builtins/pgcrypto",
         "//pkg/sql/sem/builtins/pgformat",
         "//pkg/sql/sem/cast",
         "//pkg/sql/sem/catconstants",
