--- pkg/sql/sem/builtins/pgcrypto/BUILD.bazel
@@ -1,13 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library")
-
-go_library(
-    name = "pgcrypto",
-    srcs = ["pgcrypto.go"],
-    importpath = "github.com/cockroachdb/cockroach/pkg/sql/sem/builtins/pgcrypto",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/sql/pgwire/pgcode",
-        "//pkg/sql/pgwire/pgerror",
-        "//pkg/sql/sem/eval",
-    ],
-)
